feat: refactor mission routes and add service layer for mission operations
