+++ app/models/mission.py
@@ -5,7 +5,7 @@ class MissionBase(SQLModel):
     missionname: str
 
 
-class Mission(MissionBase, SQLModel, table=True):
+class Mission(MissionBase, SQLModel, table=True):  # type: ignore
     missionid: int | None = Field(default=None, primary_key=True)
